refactor(shm): consolidate label propagation headers
