+++ kaminpar-shm/algorithms/label_propagation/node_processing.h
@@ -1,12 +1,14 @@
 /*******************************************************************************
- * Adapter between reusable iteration orders and LP node processors.
+ * Shared execution support for label-propagation node processors.
  *
- * @file:   node_processing_kernel.h
+ * @file:   node_processing.h
  * @author: Daniel Seemaier
  ******************************************************************************/
 #pragma once
 
 #include <utility>
+
+#include <tbb/enumerable_thread_specific.h>
 
 #include "kaminpar-shm/kaminpar.h"
 
@@ -14,6 +16,30 @@
 #include "kaminpar-common/random.h"
 
 namespace kaminpar::shm::lp {
+
+struct RoundStats {
+  NodeID moved = 0;
+};
+
+class RoundStatistics {
+public:
+  [[nodiscard]] RoundStats &local() {
+    return _locals.local();
+  }
+
+  [[nodiscard]] RoundStats totals() {
+    return _locals.combine([](const RoundStats lhs, const RoundStats rhs) {
+      return RoundStats{lhs.moved + rhs.moved};
+    });
+  }
+
+  void clear() {
+    _locals.clear();
+  }
+
+private:
+  tbb::enumerable_thread_specific<RoundStats> _locals;
+};
 
 enum class NodeProcessingPhase {
   SINGLE,
